refactor: DiagramEditorの階層を一つ落とす。
