+++ src/components/diagrams/Rdra20DiagramEditorContainer.vue
@@ -105,16 +105,16 @@
 
 <script lang="ts">
 import { Component, Vue, Prop, Inject, Emit } from 'nuxt-property-decorator'
-import BusinessContextDiagramEditor from '@/components/diagrams/editor/businesscontextdiagram/BusinessContextDiagramEditor.vue'
-import BusinessUseCaseDiagramEditor from '@/components/diagrams/editor/businessusecasediagram/BusinessUseCaseDiagramEditor.vue'
-import BusinessFlowDiagramEditor from '@/components/diagrams/editor/businessflowdiagram/BusinessFlowDiagramEditor.vue'
-import UsageSceneDiagramEditor from '@/components/diagrams/editor/usagescenediagram/UsageSceneDiagramEditor.vue'
-import SystemContextDiagramEditor from '@/components/diagrams/editor/systemcontextdiagram/SystemContextDiagramEditor.vue'
-import RequestModelDiagramEditor from '@/components/diagrams/editor/requrestmodel/RequestModelDiagramEditor.vue'
-import InfomationModelEditor from '@/components/diagrams/editor/infomationmodel/InfomationModelEditor.vue'
-import StateModelEditor from '@/components/diagrams/editor/statemodel/StateModelEditor.vue'
-import UseCaseCompositeDiagramEditor from '@/components/diagrams/editor/usecasecompositediagram/UseCaseCompositeDiagramEditor.vue'
-import VariationAndConditionDiagramEditor from '@/components/diagrams/editor/variationandcondition/VariationAndConditionDiagramEditor.vue'
+import BusinessContextDiagramEditor from '@/components/diagrams/editor/rdra20/businesscontextdiagram/BusinessContextDiagramEditor.vue'
+import BusinessUseCaseDiagramEditor from '@/components/diagrams/editor/rdra20/businessusecasediagram/BusinessUseCaseDiagramEditor.vue'
+import BusinessFlowDiagramEditor from '@/components/diagrams/editor/rdra20/businessflowdiagram/BusinessFlowDiagramEditor.vue'
+import UsageSceneDiagramEditor from '@/components/diagrams/editor/rdra20/usagescenediagram/UsageSceneDiagramEditor.vue'
+import SystemContextDiagramEditor from '@/components/diagrams/editor/rdra20/systemcontextdiagram/SystemContextDiagramEditor.vue'
+import RequestModelDiagramEditor from '@/components/diagrams/editor/rdra20/requrestmodel/RequestModelDiagramEditor.vue'
+import InfomationModelEditor from '@/components/diagrams/editor/rdra20/infomationmodel/InfomationModelEditor.vue'
+import StateModelEditor from '@/components/diagrams/editor/rdra20/statemodel/StateModelEditor.vue'
+import UseCaseCompositeDiagramEditor from '@/components/diagrams/editor/rdra20/usecasecompositediagram/UseCaseCompositeDiagramEditor.vue'
+import VariationAndConditionDiagramEditor from '@/components/diagrams/editor/rdra20/variationandcondition/VariationAndConditionDiagramEditor.vue'
 import Diagram from '@/domain/diagram/Diagram'
 import Resource from '@/domain/resource/Resource'
 import StorageRepository from '@/domain/storage/StorageRepository'
