+++ src/components/diagrams/editor/rdra20/infomationmodel/InfomationModelEditor.vue
